Don't load the builder overrides at compiletime.
It's already done in the code generated by run(). This also causes no plan to
be emited and TBs ending not to be executed if we don't actually run tests.
That way detecting our presence in things like Makefile.PL is possible without
messing up the output or exit code.